Adds ceph-nfs-client relation
Through the ceph-nfs-client, MicroCeph will now provision NFS clusters.

For a relation, the charm will create a NFS cluster (on up to 3 nodes),
and sets the relevant relation data for it. If there are no available
nodes, the relation will not be serviced.

The charm will take into account new nodes joining / leaving, or if a
ceph-nfs-client relation departs, provisioning nodes to the NFS clusters
as needed.

Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>